refactor: move copyFrom to S3CommonFileObject to be used by all implementations
