+++ s3-vfs/src/main/java/org/pentaho/s3common/S3CommonFileObject.java
@@ -23,10 +23,12 @@ import java.util.List;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.vfs2.FileObject;
+import org.apache.commons.vfs2.FileSelector;
 import org.apache.commons.vfs2.FileSystemException;
 import org.apache.commons.vfs2.FileType;
 import org.apache.commons.vfs2.provider.AbstractFileName;
 import org.apache.commons.vfs2.provider.AbstractFileObject;
+import org.pentaho.di.connections.vfs.provider.ConnectionFileObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -336,6 +338,56 @@ public abstract class S3CommonFileObject extends AbstractFileObject<S3CommonFile
     fileSystem.getS3Client().deleteObject( bucketName, key );
   }
 
+  /**
+   * Attempts to extract an S3FileObject from a FileObject, including wrappers like ResolvedConnectionFileObject (via reflection for getResolvedFileObject()).
+   */
+  private S3CommonFileObject extractDelegateS3FileObject( FileObject file ) {
+    if ( file instanceof S3CommonFileObject ) {
+      return (S3CommonFileObject) file;
+    }
+    if ( file instanceof ConnectionFileObject ) {
+      // If it's a ResolvedConnectionFileObject, we can try to get the underlying S3FileObject
+      ConnectionFileObject resolved = (ConnectionFileObject) file;
+      FileObject delegate = resolved.getResolvedFileObject();
+      if ( delegate instanceof S3CommonFileObject ) {
+        return (S3CommonFileObject) delegate;
+      }
+    }
+    return null;
+  }
+
+  /**
+   * Copies the content of the specified file to this file, using server-side multipart copy if both files are S3FileObjects in the same region.
+   * If the source is not an S3FileObject or is in a different region, it uses TransferManager to upload the content.
+   *
+   * @param file     The source file to copy from.
+   * @param selector A FileSelector to filter which files to copy.
+   * @throws FileSystemException If an error occurs during the copy operation.
+   */
+  @Override
+  public void copyFrom( final FileObject file, final FileSelector selector ) throws FileSystemException {
+    S3CommonFileObject s3Src = extractDelegateS3FileObject( file );
+    if ( s3Src != null ) {
+      // S3 to S3 copy
+      try {
+        logger.info( "Attempting S3→S3 server-side multipart copy from {} to {}",
+                     s3Src.getQualifiedName(), this.getQualifiedName() );
+        fileSystem.copy( s3Src, this );
+        return;
+      } catch ( FileSystemException e ) {
+        logger.warn( "S3→S3 multipart copy failed, falling back to TransferManager upload: {}", e.getMessage(), e );
+        // fallback to TransferManager upload below
+      }
+    }
+    // For non-S3FileObject or fallback, use TransferManager upload
+    try {
+      logger.info( "Uploading to S3 using TransferManager from {} to {}", file.getName(), this.getName() );
+      fileSystem.upload( file, this );
+    } catch ( Exception e ) {
+      logger.error( "TransferManager upload failed, falling back to default copy: {}", e.getMessage(), e );
+    }
+  }
+
   @Override
   public long doGetLastModifiedTime() {
     if ( s3ObjectMetadata != null && s3ObjectMetadata.getLastModified() != null ) {
@@ -431,4 +483,5 @@ public abstract class S3CommonFileObject extends AbstractFileObject<S3CommonFile
   protected String getQualifiedName( S3CommonFileObject s3nFileObject ) {
     return s3nFileObject.bucketName + "/" + s3nFileObject.key;
   }
+
 }
