+++ s3-vfs/src/test/java/org/pentaho/s3common/S3CommonFileSystemTest.java
@@ -1,0 +1,108 @@
+/*! ******************************************************************************
+ *
+ * Pentaho
+ *
+ * Copyright (C) 2024 by Hitachi Vantara, LLC : http://www.pentaho.com
+ *
+ * Use of this software is governed by the Business Source License included
+ * in the LICENSE.TXT file.
+ *
+ * Change Date: 2029-07-20
+ ******************************************************************************/
+
+
+package org.pentaho.s3common;
+
+import org.junit.Test;
+import org.mockito.Mockito;
+import org.pentaho.di.core.util.StorageUnitConverter;
+import org.pentaho.s3common.DummyS3CommonObjects.DummyS3FileObject;
+import org.pentaho.s3common.DummyS3CommonObjects.DummyS3FileSystem;
+
+import com.amazonaws.services.s3.transfer.TransferManager;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static org.pentaho.s3common.DummyS3CommonObjects.getDummyInstance;
+
+public class S3CommonFileSystemTest {
+
+  @Test
+  public void getPartSize() {
+    DummyS3FileSystem s3FileSystem = getDummyInstance();
+    s3FileSystem.storageUnitConverter = new StorageUnitConverter();
+    S3KettleProperty s3KettleProperty = mock( S3KettleProperty.class );
+    when( s3KettleProperty.getPartSize() ).thenReturn( "6MB" );
+    s3FileSystem.s3KettleProperty = s3KettleProperty;
+
+    //TEST 1: Below max
+    assertEquals( 6 * 1024 * 1024, s3FileSystem.getPartSize() );
+
+    // TEst 2: above max
+    when( s3KettleProperty.getPartSize() ).thenReturn( "600GB" );
+    assertEquals( 600L * 1024 * 1024 * 1024, s3FileSystem.getPartSize() );
+  }
+
+  @Test
+  public void testParsePartSize() {
+    DummyS3FileSystem s3FileSystem = getDummyInstance();
+    s3FileSystem.storageUnitConverter = new StorageUnitConverter();
+    long _5MBLong = 5L * 1024L * 1024L;
+    long _124MBLong = 124L * 1024L * 1024L;
+    long _5GBLong = 5L * 1024L * 1024L * 1024L;
+    long _12GBLong = 12L * 1024L * 1024L * 1024L;
+    long minimumPartSize = _5MBLong;
+    long maximumPartSize = _5GBLong;
+
+
+    // TEST 1: below minimum
+    assertEquals( minimumPartSize, s3FileSystem.parsePartSize( "1MB" ) );
+
+    // TEST 2: at minimum
+    assertEquals( minimumPartSize, s3FileSystem.parsePartSize( "5MB" ) );
+
+    // TEST 3: between minimum and maximum
+    assertEquals( _124MBLong, s3FileSystem.parsePartSize( "124MB" ) );
+
+    // TEST 4: at maximum
+    assertEquals( maximumPartSize, s3FileSystem.parsePartSize( "5GB" ) );
+
+    // TEST 5: above maximum
+    assertEquals( _12GBLong, s3FileSystem.parsePartSize( "12GB" ) );
+  }
+
+  @Test
+  public void testCopy_DelegatesToTransferManager() throws Exception {
+    DummyS3FileObject src = mock( DummyS3FileObject.class );
+    DummyS3FileObject dest = mock( DummyS3FileObject.class );
+    S3TransferManager transferManager = mock( S3TransferManager.class );
+    DummyS3FileSystem fs = Mockito.spy( getDummyInstance() );
+    doReturn( transferManager ).when( fs ).getS3TransferManager();
+    fs.copy( src, dest );
+    verify( transferManager, times( 1 ) ).copy( src, dest );
+  }
+
+  @Test
+  public void testGetS3TransferManager_CreatesWithTransferManager() {
+    DummyS3FileSystem fs = Mockito.spy( getDummyInstance() );
+    S3TransferManager transferManager = mock( S3TransferManager.class );
+    doReturn( transferManager ).when( fs ).getS3TransferManager();
+    S3TransferManager result = fs.getS3TransferManager();
+    assertNotNull( result );
+  }
+
+  @Test
+  public void testBuildTransferManager_CreatesWithS3Client() {
+    DummyS3FileSystem fs = Mockito.spy( getDummyInstance() );
+    com.amazonaws.services.s3.AmazonS3 s3Client = mock( com.amazonaws.services.s3.AmazonS3.class );
+    doReturn( s3Client ).when( fs ).getS3Client();
+    TransferManager tm = fs.buildTransferManager();
+    assertNotNull( tm );
+  }
+
+}
